feat(form): implement serializeEvent utility for form input handling
Previously, we were trying to pass the raw event, which gets serialized as `{isTrusted: true}`. This commit serializes it into the `FormEvent` type as defined in @raycast/api. It directly provides the type, target.id, and target.value. It uses a Zod schema for the value, translated directly from the FormValue_2 type in @raycast/api.