feat(footer): reorder action buttons to hide separator when highlighted
This commit moves the separator to be the last element in its parent, using `order-[n]` classes to maintain the original order. This way, we can attach "peer" classes to the buttons, allowing the separator to disappear when any peer button is hovered (`peer-hover:opacity-0`).

In the future, this may lead to weird issues with the divs not being exactly the same as the buttons. If this is the case, we may have to pass down the class. However, it should be "good enough" for now.