feat(header): move header margin from input to list
Previously, we added a bottom margin to the inputs. However, this created a weird gap, and also didn't address the padding at the bottom. Now, instead of adding this margin in the input, we move it to the `BaseList` so it can handle both while still clipping at the input.